requirements.txt
requests
pymongo
pprinter
python-dotenv
tk
loguru
deepdiff
pyinstaller
